# api/app/openapi/OpenAPI_backend_En.yaml
openapi: 3.1.0
info:
  title: Dev Portal & Translation → Chroma Pipeline API
  version: 0.3.0
  description: >
    An API that provides the pipeline: Extract → Translate → (optional Writeback) → Packaging → Chroma upsert,
    plus CRUD/import operations for the portal_* tables.
    Reflects Phase J/K policies and clarifies POST idempotency / Request-ID / 429/409 responses.
...
servers:
  - url: https://api.example.com
    description: default (switch prod/dev via env overlay)
security: []  # In the prod overlay, apply security: [{ bearerAuth: [] }]
...
paths:
...
  /translate:
    get:
      x-dev-only: true
      tags: [Translate]
      summary: List translate rows (dev)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, translated, ready_for_chroma, done, failed] }
        - in: query
          name: entity
          schema: { $ref: '#/components/schemas/Entity' }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/TranslateList' } } }
  /translate/run:
    post:
      x-internal: true
      tags: [Translate]
      summary: Translate pending → translated (Cron)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TranslateRunRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/TranslateRunResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  ############################################
  # Writeback (optional / dev-only)
  ############################################
...
  /samples/trace:
    get:
      tags: [Status]
      summary: End-to-end status per natural_key
      parameters:
        - in: query
          name: natural_key
          required: true
          schema: { $ref: '#/components/schemas/NaturalKey' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/Trace' } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Correlation ID per run. Generated by Cron/clients and attached to every request.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 256 }
      description: >
        Idempotency key for POST/PUT. Deterministically generated from a time window and payload.
        Matching retries return 200/409 (implementation-dependent) with no side effects.
  responses:
...
    ConflictIdempotent:
      description: Treated as replay due to idempotency
      headers:
        Idempotency-Status:
          description: If 'replayed', the request was treated as a replay
          schema: { type: string, enum: [replayed, conflict] }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit / temporarily refused
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: integer }
      required: [title, status]
    NaturalKey:
      type: string
      description: "Convention: entity::... e.g., field::sale.order::partner_id"
    Entity:
      type: string
      enum: [field, view_common, model, tab, smart_button, menu]
    Lang:
      type: string
      enum: [ja, en]
...
    CursorList:
      type: object
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }

    ########################################
    # Portal model
    ########################################
...
    TranslateRunRequest:
      type: object
      properties:
        limit: { type: integer, default: 200 }
        source_lang: { type: string, default: ja_JP }
        target_lang: { type: string, default: en_US }
        entities: { type: array, items: { $ref: '#/components/schemas/Entity' } }
    TranslateRunResult:
      type: object
      properties:
        picked: { type: integer }
        translated: { type: integer }
        failed: { type: integer }
        samples:
          type: array
          items: { $ref: '#/components/schemas/TranslateRow' }
    TranslateRow:
      type: object
      properties:
        natural_key: { $ref: '#/components/schemas/NaturalKey' }
        entity: { $ref: '#/components/schemas/Entity' }
        model: { type: string }
        label: { type: string }
        purpose: { type: string }
        translated_label: { type: string }
        translated_purpose: { type: string }
        status: { type: string, enum: [pending, translated, ready_for_chroma, done, failed] }
        updated_at: { type: string, format: date-time }
    TranslateList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TranslateRow' } }

    ########################################
    # Writeback
    ########################################
...
    ChromaDoc:
      type: object
      description: "doc_text is capped at 16 KiB safe length (UTF-8 safe truncation)"
      properties:
        doc_id: { type: string }
        natural_key: { $ref: '#/components/schemas/NaturalKey' }
        lang: { $ref: '#/components/schemas/Lang' }
        collection: { type: string }
        doc_text: { type: string, maxLength: 16384 }
        entity: { $ref: '#/components/schemas/Entity' }
        model: { type: string }
        model_table: { type: string }
        field_name: { type: string, nullable: true }
        action_xmlid: { type: string, nullable: true }
        target: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: true }
        status: { type: string, enum: [queued, upserted, failed] }
        updated_at: { type: string, format: date-time }
...
    Trace:
      type: object
      properties:
        natural_key: { $ref: '#/components/schemas/NaturalKey' }
        translate: { $ref: '#/components/schemas/TranslateRow' }
        chroma_docs: { type: array, items: { $ref: '#/components/schemas/ChromaDoc' } }